Create Bomo.js & overhaul index.js to use it
